Move ffconfig to root
