+++ ffconfig/main.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/hyperledger/firefly/cmd/ffconfig/migrate"
+	"github.com/hyperledger/firefly/ffconfig/migrate"
 	"github.com/spf13/cobra"
 )
